+++ changes.md
@@ -1,3 +1,14 @@
+# v0.5 - 2015-04-30
+
+- added equi7 test runner
+- fixed cell read percentage pass through to test runner
+- added tests for the test runners
+- set cells for each dataset in the test runner
+- [0] for ESA CCI
+- all cells over land for ASCAT
+- [0, 1] for Equi 7
+- installation now installes the netCDF files
+
 # v0.4 - 2015-04-29
 
 - added possibility to time each function call of the dataset functions